# indexer/services/comlink/public/swagger.json
{
  "components": {
...
    "schemas": {
      "PerpetualPositionStatus": {
        "enum": [
          "OPEN",
          "CLOSED",
          "LIQUIDATED"
        ],
        "type": "string"
      },
      "PositionSide": {
        "enum": [
          "LONG",
          "SHORT"
        ],
        "type": "string"
      },
      "IsoString": {
        "type": "string"
      },
      "PerpetualPositionResponseObject": {
        "properties": {
          "market": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/PerpetualPositionStatus"
          },
          "side": {
            "$ref": "#/components/schemas/PositionSide"
          },
          "size": {
            "type": "string"
          },
          "maxSize": {
            "type": "string"
          },
          "entryPrice": {
            "type": "string"
          },
          "realizedPnl": {
            "type": "string"
          },
          "createdAt": {
            "$ref": "#/components/schemas/IsoString"
          },
          "createdAtHeight": {
            "type": "string"
          },
          "sumOpen": {
            "type": "string"
          },
          "sumClose": {
            "type": "string"
          },
          "netFunding": {
            "type": "string"
          },
          "unrealizedPnl": {
            "type": "string"
          },
          "perpYieldIndex": {
            "type": "string"
          },
          "closedAt": {
            "allOf": [
              {
                "$ref": "#/components/schemas/IsoString"
              }
            ],
            "nullable": true
          },
          "exitPrice": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "market",
          "status",
          "side",
          "size",
          "maxSize",
          "entryPrice",
          "realizedPnl",
          "createdAt",
          "createdAtHeight",
          "sumOpen",
          "sumClose",
          "netFunding",
          "unrealizedPnl",
          "perpYieldIndex"
        ],
        "type": "object",
        "additionalProperties": false
      },
...
      "PerpetualPositionResponse": {
        "properties": {
          "positions": {
            "items": {
              "$ref": "#/components/schemas/PerpetualPositionResponseObject"
            },
            "type": "array"
          }
        },
        "required": [
          "positions"
        ],
        "type": "object",
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {}
  },
  "info": {
    "title": "Klyra Indexer API",
    "version": "v1.0.0",
    "description": "Scroll down for code samples, example requests and responses.\nBase URLs:\n* For **Testnet**, use <a href=\"https://klyra-testnet.imperator.co/v4\">https://klyra-testnet.imperator.co/v4</a>\nNote: Messages on Indexer WebSocket feeds are typically more recent than data fetched via Indexer's REST API, because the latter is backed by read replicas of the databases that feed the former. Ordinarily this difference is minimal (less than a second), but it might become prolonged under load."
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/perpetualPositions": {
      "get": {
        "operationId": "ListPositions",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerpetualPositionResponse"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "address",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "subaccountNumber",
            "required": true,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/PerpetualPositionStatus"
              }
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "createdBeforeOrAtHeight",
            "required": false,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "createdBeforeOrAt",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/IsoString"
            }
          }
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "https://klyra-testnet.imperator.co/v4",
      "description": "Testnet"
    }
  ]
}